Refactoring: Exporting functions out of operations
Refactoring:
- Operation file cleaned
- Utility files for
-- response
-- payload
-- connection
-- provider created.
