+++ src/main/java/com/mulesoft/connectors/internal/utils/ProviderUtils.java
@@ -1,0 +1,45 @@
+package com.mulesoft.connectors.internal.utils;
+
+import com.mulesoft.connectors.internal.config.InferenceConfiguration;
+
+/**
+ * Utility class for provider-specific operations.
+ */
+public class ProviderUtils {
+
+    /**
+     * Check if the inference type is OLLAMA
+     * @param configuration the connector configuration
+     * @return true if the inference type is OLLAMA, false otherwise
+     */
+    public static boolean isOllama(InferenceConfiguration configuration) {
+        return "OLLAMA".equals(configuration.getInferenceType());
+    }
+
+    /**
+     * Check if the inference type is Anthropic
+     * @param configuration the connector configuration
+     * @return true if the inference type is Anthropic, false otherwise
+     */
+    public static boolean isAnthropic(InferenceConfiguration configuration) {
+        return "ANTHROPIC".equals(configuration.getInferenceType());
+    }
+
+    /**
+     * Check if the inference type is NVIDIA
+     * @param configuration the connector configuration
+     * @return true if the inference type is NVIDIA, false otherwise
+     */
+    public static boolean isNvidia(InferenceConfiguration configuration) {
+        return "NVIDIA".equals(configuration.getInferenceType());
+    }
+
+    /**
+     * Check if the inference type is Cohere
+     * @param configuration the connector configuration
+     * @return true if the inference type is Cohere, false otherwise
+     */
+    public static boolean isCohere(InferenceConfiguration configuration) {
+        return "COHERE".equals(configuration.getInferenceType());
+    }
+}
